requirements.txt
telethon
python-dotenv
prompt_toolkit